Rename variable and change EntityListener to EventListener
