+++ src/Resources/config/services.xml
@@ -4,7 +4,7 @@
     <imports>
         <import resource="services/checker.xml"/>
         <import resource="services/controller.xml"/>
-        <import resource="services/doctrine_entity_listener.xml"/>
+        <import resource="services/event_listener.xml"/>
         <import resource="services/event_subscriber.xml"/>
         <import resource="services/fixture.xml"/>
         <import resource="services/form.xml"/>
